Wrap resnet50 model in the model factory.
